Move `parsers` to language index
