+++ src/language-css/index.js
@@ -31,8 +31,22 @@ const printers = {
   postcss: printer,
 };
 
+const parsers = {
+  // TODO: switch these to just `postcss` and use `language` instead.
+  get css() {
+    return require("./parser-postcss").parsers.css;
+  },
+  get less() {
+    return require("./parser-postcss").parsers.less;
+  },
+  get scss() {
+    return require("./parser-postcss").parsers.scss;
+  },
+};
+
 module.exports = {
   languages,
   options,
   printers,
+  parsers,
 };
