+++ src/language-handlebars/index.js
@@ -15,7 +15,14 @@ const printers = {
   glimmer: printer,
 };
 
+const parsers = {
+  get glimmer() {
+    return require("./parser-glimmer").parsers.glimmer;
+  },
+};
+
 module.exports = {
   languages,
   printers,
+  parsers,
 };
