+++ src/language-html/index.js
@@ -40,8 +40,28 @@ const printers = {
   html: printer,
 };
 
+const parsers = {
+  // HTML
+  get html() {
+    return require("./parser-html").parsers.html;
+  },
+  // Vue
+  get vue() {
+    return require("./parser-html").parsers.vue;
+  },
+  // Angular
+  get angular() {
+    return require("./parser-html").parsers.angular;
+  },
+  // Lightning Web Components
+  get lwc() {
+    return require("./parser-html").parsers.lwc;
+  },
+};
+
 module.exports = {
   languages,
   printers,
   options,
+  parsers,
 };
